requirements.txt
git+https://github.com/WHOIgit/amplify-schemas@provenance
django-ninja
